(js): extract text animation to separate file
