+++ index.html
@@ -3,6 +3,7 @@
 <body>
   <pug src="src/index.pug" />
   <script type="module" src="src/js/main.js"></script>
+  <script type="module" src="src/js/text.js"></script>
 </body>
 
 </html>
