+++ src/js/main.js
@@ -1,5 +1,4 @@
 // HTML elements
-import "../index.scss";
 
 const body = document.body;
 const toggleBtn = document.querySelector(".toggle__btn");
@@ -65,88 +64,6 @@ if (isNight() === true) {
 	}
 }
 
-// Shuffle letters
-
-class ShuffleLetters {
-	constructor(el) {
-		this.el = el;
-		this.chars = "_";
-		this.update = this.update.bind(this);
-	}
-	setText(newText) {
-		const oldText = this.el.innerText;
-		const length = Math.max(oldText.length, newText.length);
-		const promise = new Promise((resolve) => (this.resolve = resolve));
-		this.queue = [];
-
-		for (let i = 0; i < length; i++) {
-			const from = oldText[i] || "";
-			const to = newText[i] || "";
-			const start = Math.floor(Math.random() * 40);
-			const end = start + Math.floor(Math.random() * 40);
-			this.queue.push({ from, to, start, end });
-		}
-		cancelAnimationFrame(this.frameRequest);
-		this.frame = 0;
-		this.update();
-		return promise;
-	}
-	update() {
-		let output = "";
-		let complete = 0;
-		for (let i = 0, n = this.queue.length; i < n; i++) {
-			let { from, to, start, end, char } = this.queue[i];
-			if (this.frame >= end) {
-				complete++;
-				output += to;
-			} else if (this.frame >= start) {
-				if (!char || Math.random() < 0.28) {
-					char = this.randomChar();
-					this.queue[i].char = char;
-				}
-				output += `<span class="shuffled">${char}</span>`;
-			} else {
-				output += from;
-			}
-		}
-		this.el.innerHTML = output;
-		if (complete === this.queue.length) {
-			this.resolve();
-		} else {
-			this.frameRequest = requestAnimationFrame(this.update);
-			this.frame++;
-		}
-	}
-	randomChar() {
-		return this.chars[Math.floor(Math.random() * this.chars.length)];
-	}
-}
-
-function mixPhrases(css, prev, next) {
-	const phrases = [next, prev];
-
-	const el = document.querySelector(css);
-	const fx = new ShuffleLetters(el);
-
-	let counter = 0;
-	const callback = () => {
-		fx.setText(phrases[counter]);
-		counter = (counter + 1) % phrases.length;
-	};
-	return callback;
-}
-
-const first = mixPhrases(".title__dynamic-first", "I", "X");
-const second = mixPhrases(".title__dynamic-second", "velop", "sign");
-
-if (first.length && second.length) {
-	const result = () => {
-		first();
-		second();
-	};
-	setInterval(result, 4500);
-}
-
 // @LM img src error handler, will be utilized in the future
 async function checkImage(url) {
 	try {
@@ -171,10 +88,10 @@ const cardTarget = document.querySelector(".card");
 const scrollTarget = document.querySelector(".footer__scroll-top button");
 
 // @LM buggy solution
-cardTarget.addEventListener("click", function (e) {
-	folioTarget.scrollIntoView({ behavior: "smooth" });
-});
+// cardTarget.addEventListener("click", function (e) {
+// 	folioTarget.scrollIntoView({ behavior: "smooth" });
+// });
 
-scrollTarget.addEventListener("click", function (e) {
-	folioTarget.scrollIntoView({ behavior: "smooth" });
-});
+// scrollTarget.addEventListener("click", function (e) {
+// 	folioTarget.scrollIntoView({ behavior: "smooth" });
+// });
